refactor(chat): split streaming pipeline modules
- Extract SSE persistence, Redis run lifecycle, and agent generation into streaming modules
- Add thin StreamChatHandler wrapper around chat stream setup
- Preserve provider stream_error metadata and cover it with an integration test
- Update chat stream lifecycle tests for new module boundaries
